Split scenarios service.
