+++ raiden/scenario_player/services/runner/blueprints/__init__.py
@@ -1,0 +1,1 @@
+from raiden.scenario_player.services.runner.blueprints.runner import runner_view
