--- raiden/scenario_player/services/scenario/blueprint/__init__.py
@@ -1,3 +1,0 @@
-from raiden.scenario_player.services.scenario.blueprint.scenario import scenario_view
-from raiden.scenario_player.services.scenario.blueprint.runner import runner_view
-from raiden.scenario_player.services.scenario.blueprint.validate import validate_view
