+++ raiden/scenario_player/services/scenario/blueprints/__init__.py
@@ -1,0 +1,3 @@
+from raiden.scenario_player.services.scenario.blueprints.scenario import scenario_view
+from raiden.scenario_player.services.scenario.blueprints.runner import runner_view
+from raiden.scenario_player.services.scenario.blueprints.validate import validate_view
