+++ raiden/scenario_player/services/scenario/blueprints/validate.py
